src/requirements.txt
attrs>=23.2.0
cachetools>=5.3.3
httpx>=0.23.0
pytest>=6.2.5
authlib>=1.6.0
requests>=2.31.0
fastavro>=1.11.0
fastapi>=0.95.0
uvicorn>=0.21.0
confluent-kafka>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0